User page (#88)
* created user page

* Created the user page and load edit stuff based on
session in store.

* Created a form that correctly displays the list
info when opened. Opens in a model. Does not do anything yet but will
in the next commit

* Completed list form except for list tags
list tags requires an api change in the model so for right
now that feature won't work. Need to do update and delete for
experiences and descriptions

* Bug Fix

Fixed 2 bugs
- Sometimes when editing a list, the list's details wouldn't show,
this was due to the form resetting the state of the form when it closed.
The fix is only temporary as I believe that if I was to edit the state,
then cancel, it won't reset back to the original state
- When adding a list, it wouldn't associate with the user. This was due
to the userId not being used when adding the list to the database

* Bug Fixes
- Fixed a bug where when the edit modal for a list closed without
actually updating the list, it wouldn't reset the data back to the
original data. This was due to the data not being manualy reset. It is
now
- Added a section after the edit dispatch has been completed to update
the state vars so that if a user choses to edit the form again, the
correct values will be present

* Removed unused console log and added signout
functionality

* Fixed sign up
NOTE that I need to get server errors so I can let the user know
of any other failures, like the username/email existing or username
being too long.

* Fixed some bugs

* Experience and Description forms
Added in the experience and description forms. They will reqire some
more testing.

* Added update functionality to desc and exp

* Added delete functionality to desc and exp

* styling changes

* updated auth forms

* List tag update

Updated some styles and added a search functionality to the lists
tab under the user page